convert datadog_track_errors to metrics_track_errors
