+++ corehq/apps/export/export.py
@@ -6,7 +6,7 @@ from collections import Counter
 
 from couchdbkit import ResourceConflict
 
-from corehq.util.metrics import metrics_counter
+from corehq.util.metrics import metrics_counter, metrics_track_errors
 from couchexport.export import FormattedRow, get_writer
 from couchexport.models import Format
 from dimagi.utils.logging import notify_exception
@@ -26,7 +26,6 @@ from corehq.apps.export.models.new import (
 )
 from corehq.elastic import iter_es_docs_from_query
 from corehq.toggles import PAGINATED_EXPORTS
-from corehq.util.datadog.gauges import datadog_track_errors
 from corehq.util.datadog.utils import DAY_SCALE_TIME_BUCKETS, load_counter
 from corehq.util.files import TransientTempfile, safe_filename
 
@@ -435,7 +434,7 @@ def _get_base_query(export_instance):
         )
 
 
-@datadog_track_errors('rebuild_export', duration_buckets=DAY_SCALE_TIME_BUCKETS)
+@metrics_track_errors('rebuild_export')
 def rebuild_export(export_instance, progress_tracker):
     """
     Rebuild the given daily saved ExportInstance
